requirements.txt
PyYAML<6,>=5.1
